Implement custom pipeline for interactive kw plots
This is a monster commit containing a lot of customisations to cluster
keyword embeddings per register, with custom label mappings and hover
text fields for better interpretability in the interactive plots. This
version was used to run publication-ready interactive plots, save for
a review step for the cluster label mapping, which is as of now pending.